+++ internal/src/main/java/com/affise/attribution/internal/callback/InternalResult.kt
@@ -1,6 +1,6 @@
 package com.affise.attribution.internal.callback
 
-interface AffiseResult {
+interface InternalResult {
     fun success(data: Any?)
     fun error(error: String)
     fun notImplemented()