+++ internal/src/main/java/com/affise/attribution/internal/data/DataName.kt
@@ -1,0 +1,41 @@
+package com.affise.attribution.internal.data
+
+internal object DataName {
+    const val SUCCESS = "success"
+    const val ERROR = "error"
+
+    const val PRODUCTS = "products"
+    const val INVALID_IDS = "invalidIds"
+    const val PRODUCT = "product"
+    const val ORDER_ID = "orderId"
+    const val ORIGINAL_ORDER_ID = "originalOrderId"
+    const val PURCHASE = "purchase"
+
+    const val DEEPLINK = "deeplink"
+    const val SCHEME = "scheme"
+    const val HOST = "host"
+    const val PATH = "path"
+    const val PARAMETERS = "parameters"
+
+    const val METHOD = "method"
+    const val URL = "url"
+    const val CODE = "code"
+    const val MESSAGE = "message"
+    const val HEADERS = "headers"
+    const val BODY = "body"
+
+    const val PRODUCT_ID = "productId"
+    const val TITLE = "title"
+    const val DESCRIPTION = "description"
+    const val PRODUCT_TYPE = "productType"
+    const val PRICE = "price"
+    const val VALUE = "value"
+    const val CURRENCY_CODE = "currencyCode"
+    const val FORMATTED_PRICE = "formattedPrice"
+    const val SUBSCRIPTION = "subscription"
+    const val OFFER_TOKEN = "offerToken"
+    const val OFFER_ID = "offerId"
+    const val TIME_UNIT = "timeUnit"
+    const val NUMBER_OF_UNITS = "numberOfUnits"
+    const val PRODUCT_DETAILS = "productDetails"
+}